requirements.txt
# Dependencias del proyecto POS HTF
PySide6>=6.6.0  # Framework de UI moderno con Qt6
QtAwesome>=1.4.0  # Iconos FontAwesome para Qt
python-dotenv>=1.0.0
supabase>=2.0.0  # Cliente para sincronización con Supabase
psycopg2-binary>=2.9.0  # Driver PostgreSQL
pyinstaller>=6.0.0  # Para generar ejecutable de Windows